Refactors free tile component out of free tile row so useDraggable hook isn't missing on re-render
